% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sector.R
\name{assign_sector}
\alias{assign_sector}
\title{assign_sector}
\usage{
assign_sector(data = NULL)
}
\arguments{
\item{data}{A dataframe with a fleet code column (sflt_desc_id)}
}
\value{
a data.frame with a new column for fleet/sector
}
\description{
Assign fleet sectors to data.
}
\examples{
\dontrun{
example <- assign_sector(df)
}
}
