% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_zone.R
\name{assign_zone}
\alias{assign_zone}
\title{assign_zone}
\usage{
assign_zone(
  marfis.df = NULL,
  isdb.df = NULL,
  y = year,
  lat.field = "lat",
  lon.field = "lon",
  dir = "C:/LocalDataDump/Fleets",
  extract = "F"
)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with latitude and longitude columns.}

\item{isdb.df}{A dataframe of ISDB data for the relevant year. Default is "isdb" as supplied by the function isdbpull.R}

\item{y}{Year}

\item{lat.field}{Latitude field, in decimal degrees. Default is "lat".}

\item{lon.field}{Longitude field, in decimal degrees. Default in "lon".}

\item{dir}{Directory with locally stored data, which can be generated with the function localcopies.R. Default is "C:/LocalDataDump/Fleets"}

\item{extract}{Choose whether or not to force extract all local copies of required MARFIS and ISDB tables (T/F). Default is "F", which extracts only missing tables and does not update existing ones.}
}
\value{
a list of two dataframes: one with an additional column for Zone, with all Zones assigned; another with an additional column for Zone, with no Zones assigned.
}
\description{
Assign discard zones to data.
}
\examples{
\dontrun{
example1 <- assign_zone(marfis)
}
}
