% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codforbid2-package.R
\docType{package}
\name{codforbid2-package}
\alias{codforbid2}
\alias{codforbid2-package}
\title{codforbid2: Georges Bank Cod Discard Calculation}
\description{
Functions to replicate the cod discard calculation for Georges Bank, according to the method from Gavaris et al. (2007).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CMClark1/codforbid2}
  \item Report bugs at \url{https://github.com/CMClark1/codforbid2/issues}
}

}
\author{
\strong{Maintainer}: Caira Clark \email{caira.clark@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-1111-2222-3333}{ORCID})

}
\keyword{internal}
