% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_summary.R
\name{coverage_summary}
\alias{coverage_summary}
\title{coverage_summary}
\usage{
coverage_summary(marfis.df = NULL, y = year)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with observer TRIP ID codes.}

\item{y}{Year}
}
\value{
An Excel document (Discards_ObsCoverageSummary_2024.xlsx) of observer coverage in the working directory.
}
\description{
Export an observer coverage summary Excel document.
}
\examples{
\dontrun{
example <- coveragesummary(marfis.df = marfis, y = 2023)
}
}
