% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egb_qaqc.R
\name{egb_qaqc}
\alias{egb_qaqc}
\title{egb_qaqc}
\usage{
egb_qaqc(
  marfis.df = NULL,
  isdb.df = NULL,
  savedoutput = "T",
  directory = getwd()
)
}
\arguments{
\item{marfis.df}{The name of the MARFIS dataframe previously pulled using the marfispull function.}

\item{isdb.df}{The name of the ISDB dataframe previously pulled using the isdbpull function.}

\item{savedoutput}{Specify whether or not you could like the output files saved in the working directory for review or action (see below for the descriptions of these files). Default is "T" (true).}

\item{directory}{Directory where output files will be saved. Default is the working directory.}
}
\value{
Returns a data frame of QAQC's data ready for further cleaning, and five Excel or .csv files for review or action: Discards_ISDB_check to send to the observer companies, Discards_MARFIS_missing to send to CDD, an observer coverage summary (Discards_ObsCoverageSummary), removed records (Discards_RemovedRecords), aggregated data (Discards_MARFISXtab_Aggregated) and data for grouping (Discards_DataforGrouping).
}
\description{
A function to QAQC MARFIS data using the ISDB database.
}
\examples{
\dontrun{
example1 <- egb_qaqc(marfis.df=marfis, isdb.df=isdb)
}
}
