% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdbpull.R
\name{isdbpull}
\alias{isdbpull}
\title{isdbpull}
\usage{
isdbpull(
  year = NULL,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn
)
}
\arguments{
\item{year}{The year required.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}
}
\value{
a data.frame of observed trips
}
\description{
Pull observed trips for the groundfish fleet from the ISDB database.
}
\examples{
\dontrun{
example1 <- isdbpull(year = 2024)
}
}
