% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfispull.R
\name{marfispull}
\alias{marfispull}
\title{marfispull}
\usage{
marfispull(
  year = NULL,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn
)
}
\arguments{
\item{year}{The year required.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}
}
\value{
a data.frame of landings
}
\description{
Pull landings data for the groundfish fleet from DFO's Maritime Fishery Information System (MARFIS) database.
}
\examples{
\dontrun{example <- marfispull(year=2024)}
}
