% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_panel.R
\name{no_panel}
\alias{no_panel}
\title{no_panel}
\usage{
no_panel(
  marfis.df = NULL,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn,
  y = year
)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with TRIP codes.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}

\item{y}{Year}
}
\value{
a list of two dataframes: one with trips that used a separator panel to keep; another with trips that did not use a separator panel to remove
}
\description{
Remove trips that did not use a separator panel
}
\examples{
\dontrun{
example1 <- nopanel(marfis.df = marfis)
}
}
