% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_sought.R
\name{species_sought}
\alias{species_sought}
\title{species_sought}
\usage{
species_sought(
  marfis.df = NULL,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn,
  y = year
)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with TRIP codes.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}

\item{y}{Year}
}
\value{
a list of two data.frames: one with trips that sought haddock to keep; another with trips that sought other species to remove.
}
\description{
Remove trips that sought species other than haddock, and then remove
}
\examples{
\dontrun{
example <- species_sought(marfis.df=data)
}
}
